+++ src/Core/CommonCQRSTrait.php
@@ -98,33 +98,62 @@ trait CommonCQRSTrait
             $this->lastUsedDecoders[$profilerKey] = [];
         }
 
-        // $dump = [
-        //     'initiator' => Utils::getType($initiator),
-        // ];
+        $this->verbose($currentProfileKey, fn () => [
+            'start decoding response' => Utils::getType($currentResponse),
+        ]);
 
         $i = -1;
         foreach ($this->decoders as $decoderItem) {
             $i++;
             $decoder = $decoderItem->getItem();
-            //$dump['loop_' . $i] = [];
-            //$dump['loop_' . $i]['trying decoder'] = Utils::getType($decoder);
+
+            $this->debug($currentProfileKey, fn () => [
+                'loop' => $i,
+                'trying decoder' => Utils::getType($decoder),
+            ]);
 
             if ($decoder->supports($currentResponse, $initiator)) {
-                //$dump['loop_' . $i]['decoder supports response'] = $currentResponse;
+                $this->debug($currentProfileKey, fn () => [
+                    'loop' => $i,
+                    'decoder' => Utils::getType($decoder),
+                    'supports response' => Utils::getType($currentResponse),
+                ]);
+
                 $decodedResponse = $this->getDecodedResponse(
                     $initiator,
                     $currentProfileKey,
                     $decoder,
                     $currentResponse
                 );
-                //$dump['loop_' . $i]['decoder decoded response'] = $decodedResponse;
 
-                //if (function_exists('dump')) {
-                //    call_user_func('dump', [__METHOD__ => $dump]);
-                //}
+                $this->verboseOrDebug(
+                    $currentProfileKey,
+                    fn () => [
+                        'loop' => $i,
+                        'decoder' => Utils::getType($decoder),
+                        'response' => Utils::getType($currentResponse),
+                        'decoded response' => Utils::getType($decodedResponse),
+                    ],
+                    fn () => [
+                        'loop' => $i,
+                        'decoder' => Utils::getType($decoder),
+                        'response' => [
+                            'type' => Utils::getType($currentResponse),
+                            'data' => $currentResponse,
+                        ],
+                        'decoded response' => [
+                            'type' => Utils::getType($decodedResponse),
+                            'data' => $decodedResponse,
+                        ],
+                    ]
+                );
 
                 if ($decodedResponse instanceof DecodedValueInterface) {
                     $decodedResponse = $decodedResponse->getValue();
+
+                    $this->verbose($currentProfileKey, fn () => [
+                        'decoding is finished' => sprintf('DecodedValue<%s>', Utils::getType($decodedResponse)),
+                    ]);
 
                     if ($profilerKey) {
                         $this->lastUsedDecoders[$profilerKey][] = sprintf(
@@ -148,13 +177,41 @@ trait CommonCQRSTrait
                     );
                 }
                 $currentResponse = $decodedResponse;
-                //} else {
-                //if (function_exists('dump')) {
-                //    call_user_func('dump', [__METHOD__ => $dump]);
-                //}
             }
         }
 
         $this->lastSuccess = $currentResponse;
+    }
+
+    private function verboseOrDebug(?UuidInterface $profilerKey, callable $verboseData, callable $debugData): void
+    {
+        if ($this->profilerBag
+            && $profilerKey !== null
+            && ($profilerItem = $this->profilerBag->get($profilerKey))
+        ) {
+            // todo - it could be better to add a specific array for verbose and debug to the profilerItem, but to gather the info and test it, this should be enough
+
+            if ($this->profilerBag->isDebug()) {
+                $debug = $profilerItem->getAdditionalData()['debug'] ?? [];
+                $debug[] = $debugData();
+                $profilerItem->setAdditionalData('debug', $debug);
+            } elseif ($this->profilerBag->isVerbose()) {
+                $verbose = $profilerItem->getAdditionalData()['verbose'] ?? [];
+                $verbose[] = $verboseData();
+                $profilerItem->setAdditionalData('verbose', $verbose);
+            }
+        }
+    }
+
+    /** @phpstan-param callable(): array $data */
+    private function verbose(?UuidInterface $profilerKey, callable $data): void
+    {
+        $this->verboseOrDebug($profilerKey, $data, fn () => []);
+    }
+
+    /** @phpstan-param callable(): array $data */
+    private function debug(?UuidInterface $profilerKey, callable $data): void
+    {
+        $this->verboseOrDebug($profilerKey, fn () => [], $data);
     }
 }